refacto: refacto du form edition-rdv
